Fixed global_ses_test

+++ openquake/hazardlib/calc/stochastic.py
@@ -34,7 +34,7 @@ from openquake.hazardlib.countries import MODELS
 TWO16 = 2 ** 16  # 65,536
 TWO30 = 2 ** 30  # 1,073,741,824
 TWO32 = 2 ** 32  # 4,294,967,296
-TWO60 = 2 ** 60
+TWO60 = numpy.int64(2 ** 60)
 F64 = numpy.float64
 U16 = numpy.uint16
 U32 = numpy.uint32
